move generic toast consumers out of components
