+++ src/toasts/AnalyticsToast.tsx
@@ -1,27 +1,29 @@
 /*
-Copyright 2020 The Matrix.org Foundation C.I.C.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
+ *
+ * Copyright 2019 Michael Telatynski <7t3chguy@gmail.com>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * /
+ */
 
 import React from "react";
 
-import {_t} from "../../../languageHandler";
-import ToastStore, {Priority} from "../../../stores/ToastStore";
-import dis from "../../../dispatcher/dispatcher";
-import Analytics from "../../../Analytics";
-import AccessibleButton from "../elements/AccessibleButton";
-import GenericToast from "./GenericToast";
+import {_t} from "../languageHandler";
+import ToastStore, {Priority} from "../stores/ToastStore";
+import dis from "../dispatcher/dispatcher";
+import Analytics from "../Analytics";
+import AccessibleButton from "../components/views/elements/AccessibleButton";
+import GenericToast from "../components/views/toasts/GenericToast";
 
 const onAccept = () => {
     dis.dispatch({
