+++ src/toasts/ServerLimitToast.tsx
@@ -1,25 +1,27 @@
 /*
-Copyright 2020 The Matrix.org Foundation C.I.C.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
+ *
+ * Copyright 2019 Michael Telatynski <7t3chguy@gmail.com>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * /
+ */
 
 import React from "react";
 
-import {_t, _td} from "../../../languageHandler";
-import ToastStore, {Priority} from "../../../stores/ToastStore";
-import GenericToast from "./GenericToast";
-import {messageForResourceLimitError} from "../../../utils/ErrorUtils";
+import {_t, _td} from "../languageHandler";
+import ToastStore, {Priority} from "../stores/ToastStore";
+import GenericToast from "../components/views/toasts/GenericToast";
+import {messageForResourceLimitError} from "../utils/ErrorUtils";
 
 const TOAST_KEY = "server-limit";
